+++ appData/src/gb/src/core/vm.c
@@ -122,9 +122,9 @@ void vm_jump(SCRIPT_CTX * THIS, UBYTE * pc) OLDCALL BANKED {
 
 UBYTE wait_frames(void * THIS, UBYTE start, UWORD * stack_frame) OLDCALL BANKED {
     // we allocate one local variable (just write ahead of VM stack pointer, we have no interrupts, our local variables won't get spoiled)
-    if (start) *((SCRIPT_CTX *)THIS)->stack_ptr = sys_time;
+    if (start) *((SCRIPT_CTX *)THIS)->stack_ptr = stack_frame[0] + 1; // Store the number of frames to wait
     // check wait condition
-    return (((UWORD)sys_time - *((SCRIPT_CTX *)THIS)->stack_ptr) < stack_frame[0]) ? ((SCRIPT_CTX *)THIS)->waitable = TRUE, (UBYTE)FALSE : (UBYTE)TRUE;
+    return ((--*((SCRIPT_CTX *)THIS)->stack_ptr) != 0) ? ((SCRIPT_CTX *)THIS)->waitable = TRUE, (UBYTE)FALSE : (UBYTE)TRUE;
 }
 // calls C handler until it returns true; callee cleanups stack
 void vm_invoke(SCRIPT_CTX * THIS, UBYTE bank, UBYTE * fn, UBYTE nparams, INT16 idx) OLDCALL BANKED {